requirements.txt
psutil>=5.9.0
google-generativeai>=0.8.0
rich>=13.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
